requirements.txt
Django==3.2.18
psycopg2-binary==2.9.5
python-dotenv==1.0.0
djangorestframework==3.14.0
coverage==7.2.1